Create Unit3Lab.py
Unit 3 lab: On a roulette wheel, the pockets are numbered from 0 to 36. The colors of the pockets are as follows:

Pocket 0 is green.
For pockets 1 through 10
The odd-numbered pockets are red
The even- numbered pockets are black
For pockets 11 through 18
The odd-numbered pockets are black
The even- numbered pockets are red
For pockets 19 through 28
The odd-numbered pockets are red
The even- numbered pockets are black
For pockets 29 through 36
The odd-numbered pockets are black
The even- numbered pockets are red
Write a program that asks the user to enter a pocket number and displays whether the pocket is green, red, or black.